+++ src/main/java/com/zjut/smartClassroom/Service/impl/ProblemSetServiceImpl.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
-public class ProblemSetImpl implements ProblemSetService {
+public class ProblemSetServiceImpl implements ProblemSetService {
 
     /**
      *@author John
