+++ src/main/java/com/zjut/smartClassroom/repository/ProblemSetRepository.java
@@ -3,7 +3,10 @@ package com.zjut.smartClassroom.repository;
 import com.zjut.smartClassroom.dataObject.ProblemSet;
 import com.zjut.smartClassroom.dataObject.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Date;
 
@@ -32,6 +35,6 @@ public interface ProblemSetRepository extends JpaRepository<ProblemSet, Integer>
     // 根据练习集id修改练习集
     @Modifying
     @Transactional
-    @Query("update ProblemSetPart p set p.problemSetName=?1,p.problemSetDetail=?2,p.problemReleaseTime=?3,p.problemSetPicUrl=?4,p.paperId=?5 where p.problemSetId = ?6")
+    @Query("update ProblemSet p set p.problemSetName=?1,p.problemSetDetail=?2,p.problemReleaseTime=?3,p.problemSetPicUrl=?4,p.paperId=?5 where p.problemSetId = ?6")
     int updateDataByProblemSetId(String problemSetName, String problemSetDetail, Date problemReleaseTime, String problemSetPicUrl, int paperId, int problemSetId);
 }
